:truck: Move `CompilationUnit` into separate file
